third_party/WebKit/LayoutTests/media/video-controls-download-button-not-displayed-mediastream.html

<!DOCTYPE HTML>
<title>Download button should not appear for MediaStream videos.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<video controls></video>
<script>

async_test(function(t) {
  var v = document.querySelector('video');
  v.srcObject = (new AudioContext()).createMediaStreamDestination().stream;

  var controlID = "-internal-download-button";
  var downloadButton = mediaControlsElement(internals.shadowRoot(v).firstChild, controlID);
  setTimeout(t.step_func_done(function() {
    assert_equals(getComputedStyle(downloadButton).display, "none");
    v.srcObject = null;
  }), 0);
});
</script>
